requirements.txt
flask
numpy
requests
python-dotenv
openpyxl
gunicorn